Adicionado hooks useAPI para pesquisar
